fix remove answer

+++ app/controllers/journal_entries_controller.rb
@@ -47,7 +47,7 @@ class JournalEntriesController < ApplicationController # < ActiveRbac::Component
     entry = JournalEntry.find(params[:id])
     # remove any score report created
     # delete all answers and answer cells, delete login for journal_entry
-    entry.destroy
+    entry.destroy if !entry.nil?
 
     if entry && entry.survey_answer
       sc = ScoreRapport.find_by_survey_answer_id(entry.survey_answer.id)
